+++ src/titiler/application/titiler/application/__init__.py
@@ -1,3 +1,3 @@
 """titiler.application"""
 
-from .version import __version__  # noqa
+__version__ = "0.5.1"