+++ src/titiler/core/titiler/core/__init__.py
@@ -1,10 +1,11 @@
 """titiler.core"""
 
-from . import dependencies, errors, factory, routing, utils, version  # noqa
+__version__ = "0.5.1"
+
+from . import dependencies, errors, factory, routing, utils  # noqa
 from .factory import (  # noqa
     BaseTilerFactory,
     MultiBandTilerFactory,
     MultiBaseTilerFactory,
     TilerFactory,
 )
-from .version import __version__  # noqa